Move Redis subcultures into redis module
